requirements.txt
altgraph==0.17.3
Babel==2.11.0
customtkinter==5.1.2
darkdetect==0.8.0
fpdf==1.7.2
greenlet==2.0.2
pefile==2023.2.7
Pillow==9.4.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.0
SQLAlchemy==2.0.4
tkcalendar==1.6.1
typing_extensions==4.5.0
